simplified and finalized background service
